third_party/blink/web_tests/platform/generic/fast/parser/iframe-sets-parent-to-javascript-url-expected.txt

ALERT: 1
ALERT: 2
ALERT: 3
ALERT: 4


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
